Made some changes to support per-host differences.

+++ .bash/aliases.sh
@@ -1,7 +1,8 @@
 # OS/Apps
 # =======
-alias ..='cd ..'     # Go up one directory
-alias ...='cd ../..' # Go up two directories
+alias ..='cd ..'         # Go up 1 directory
+alias ...='cd ../..'     # Go up 2 directories
+alias ....='cd ../../..' # Go up 3 directories
 
 alias l='ls -lah'    # Long view, show hidden
 alias la='ls -AF'    # Compact view, show hidden
@@ -14,7 +15,8 @@ alias grepsvn='grep -r --exclude-dir="\.svn"'
 # Python/Django
 alias pypath='python -c "import sys; print sys.path" | tr "," "\n" | grep -v "egg"'
 
-# Mac OS X
-alias hiddenshow='defaults write com.apple.finder AppleShowAllFiles -bool true; KillAll Finder;'
-alias hiddenhide='defaults write com.apple.finder AppleShowAllFiles -bool false; KillAll Finder;'
-alias changes='rm ~/Library/Application\ Support/Changes/.cacheSettings2'
+if [ $HOST_TYPE = "mac" ]; then
+    alias hiddenshow='defaults write com.apple.finder AppleShowAllFiles -bool true; KillAll Finder;'
+    alias hiddenhide='defaults write com.apple.finder AppleShowAllFiles -bool false; KillAll Finder;'
+    alias changes='rm ~/Library/Application\ Support/Changes/.cacheSettings2'
+fi
